# apps/backend/openapi.test3.yaml
openapi: 3.0.3
info:
  title: X-Ear CRM API (auto-generated)
  version: 0.0.0-auto
servers:
- url: http://localhost:5003
paths:
...
  /api/apps:
    get:
      summary: GET /api/apps
      operationId: apps_list_apps
      tags:
      - Apps
      responses:
        '200':
          description: Success
    post:
      summary: POST /api/apps
      operationId: apps_create_app
      tags:
      - Apps
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/apps/{app_id}:
    get:
      summary: GET /api/apps/{app_id}
      operationId: apps_get_app
      tags:
      - Apps
      parameters:
      - name: app_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
    put:
      summary: PUT /api/apps/{app_id}
      operationId: apps_update_app
      tags:
      - Apps
      parameters:
      - name: app_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
    delete:
      summary: DELETE /api/apps/{app_id}
      operationId: apps_delete_app
      tags:
      - Apps
      parameters:
      - name: app_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
  /api/apps/{app_id}/assign:
    post:
      summary: POST /api/apps/{app_id}/assign
      operationId: apps_assign_role_to_user
      tags:
      - Apps
      parameters:
      - name: app_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/apps/{app_id}/transfer_ownership:
    post:
      summary: POST /api/apps/{app_id}/transfer_ownership
      operationId: apps_transfer_ownership
      tags:
      - Apps
      parameters:
      - name: app_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
...
  /api/campaigns:
    get:
      summary: GET /api/campaigns
      operationId: campaigns_get_campaigns
      tags:
      - Campaigns
      responses:
        '200':
          description: Success
    post:
      summary: POST /api/campaigns
      operationId: campaigns_create_campaign
      tags:
      - Campaigns
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/campaigns/{campaign_id}/send:
    post:
      summary: POST /api/campaigns/{campaign_id}/send
      operationId: campaigns_send_campaign
      tags:
      - Campaigns
      parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/cash-records:
    get:
      summary: Return cash register records for cashflow.html consumption.
      description: 'Maps PaymentRecord rows to a simplified format:

        {

        id, date, transactionType, recordType, patientId, patientName, amount, description

        }

        Optional query params:

        - limit: max number of rows (default 200)

        - start_date, end_date: ISO date filters

        - status: filter by payment status (paid/pending/partial)'
      operationId: cash_records_get_cash_records
      tags:
      - Cash-records
      responses:
        '200':
          description: Success
    post:
      summary: Create a new cash record from cashflow.html form submission.
      operationId: cash_records_create_cash_record
      tags:
      - Cash-records
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/cash-records/{record_id}:
    delete:
      summary: Delete a cash record by ID.
      operationId: cash_records_delete_cash_record
      tags:
      - Cash-records
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
...